refactor(Questionnaire): Fix file structure
