+++ src/pages/questionnaires/index.tsx
@@ -4,8 +4,8 @@ import { useRouter } from "next/router";
 import React from "react";
 import { appBaseStyle, transition } from "../../components/layout";
 import { PageHeader } from "../../components/PageHeader";
-import { QuestionnaireListContainer } from "../../components/QuestionnaireListContainer";
-import { QuestionnaireConnectionDocument, QuestionnaireConnectionQuery, QuestionnaireConnectionQueryVariables } from "../../components/QuestionnaireListContainer/__generated__/index.generated";
+import { QuestionnaireListContainer } from "../../components/Questionnaire";
+import { QuestionnaireConnectionDocument, QuestionnaireConnectionQuery, QuestionnaireConnectionQueryVariables } from "../../components/Questionnaire/__generated__/QuestionnaireListContainer.generated";
 import { addApolloStateToPageProps, initializeApollo } from "../../hooks/useAplloClient";
 import { PrimaryButton } from "../../lib/components/Button";
